Merge r1801144, r1801148, r1801456 from trunk:
mpm_winnt: Factor out a helper function to parse the type of an accept
filter and use an appropriate enum for it.

This makes the code in winnt_accept() a bit easier to follow.  As a minor
side effect, it also fixes a small bug where the "unrecognized AcceptFilter
'%s'" log entry would always contain "none" instead of the actually
unrecognized kind of the accept filter.


mpm_winnt: Fix typo in the logged message in winnt_get_connection().


mpm_winnt: Following up on r1801144, use the new accept_filter_e enum
values in a couple of missed places in winnt_accept().

Submitted by: kotkov
Reviewed by: jailletc36, jim (via inspection), wrowe


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/2.4.x@1840546 13f79535-47bb-0310-9956-ffa450edef68