Quireference extract the default value of a directive  from the <default> of the corresponding <directivesynopsis> block.
The current implementation  expects <default> to be formated like:
directive_name default_value

If the directive name is not there, no default value can be extracted and it is missing in the quickreference table.

This commit adds a basic check to see if the formating looks correct. It catches cases where the directive name is  omited or is misspelled.

If it generates too much noise, feel free to revert or improve :)

git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@1852285 13f79535-47bb-0310-9956-ffa450edef68