mod_rewrite.xml: add an example about override behavior
A user in PR 63051 was confused about the default behavior
of mod_rewrite (override) when merging two sections in the
directory context at the same nesting level. I was confused
as well, but after a quick look at the code I figured out that
RewriteOptions Inherit seems to solve the issue (naming is,
in this case, a bit counter intuitive). I hope that I added
a good clarification in the code, if not please amend it as
you think it's best!


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@1853280 13f79535-47bb-0310-9956-ffa450edef68