s/cacheing/caching/ Improve layout.
Add some missing <module> tags.
Add some missing <directive> tags.
Add some missing <directive module=...> links
Add the name of the directive in <default>, so that 'quickreference.html' is generated correctly.
In the AuthnCacheContext directive, fix the way 'directory' and 'server' are displayed. They are key-words and should not be in italic.

git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@1865402 13f79535-47bb-0310-9956-ffa450edef68