Merge r1872394 from trunk
   * In winnt_accept() (server/mpm/winnt/child.c), there's a call to PostQueuedCompletionStatus() with a buffer (BytesRead) for the number of bytes read.

When compiling the code with Visual Studio 2019 (v 16.3.10), we get a warning than BytesRead is not initialized. When using /RTCu (run-time checks for uninitialzed variables), the program crashes on this line with an exception.

If we initialize the variable to 0, the problem is solved.

PR: 63965
Submitted by: gbechis
Reviewed by: gbechis,jailletc36 (+ CTR rule for non-Unix, single-platform code)
Backported by: jailletc36

git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/2.4.x@1874971 13f79535-47bb-0310-9956-ffa450edef68