[SECURITY] Don't stop Certificate Revoked messages.
Certificate Revoked Responder messages don't belong to 'error' class.
When the server receives one, it MUST be passed on to the client.
And stored for the normal period of basic responses.

Also don't log an error each time it is retrieved from cache,
only once when it is retrieved from the OCSP responder.