Process early errors via a dummy HTTP/1.1 request as well
Process early errors via a dummy HTTP/1.1 request as well to ensure
that the request gets logged correctly and possible custom error
pages are considered. The previous way of directly sending a HTTP/2
answer with the HTTP status code appropriate for the error is more
efficient, but does not log the request nor sents a possible custom
error page.

* modules/http2/h2.h: Add early_http_status to h2_request struct.

* modules/http2/h2_request.c(h2_request_create_rec): Check if
  early_http_status is set for the request and die with the
  status code it contains if set.

* modules/http2/h2_session.c(on_header_cb): Adjust the error condition
  now that we mark early errors via early_http_status: Only return an
  error if the status is not success and don't have early_http_status
  set.

* modules/http2/h2_stream.c(set_error_response): Set early_http_status
  on the request instead of creating headers for a response and a
  respective brigade.