Merge r1885939, r1885940, r1885941, r1885945 from trunk:
Do not allow to set empty bind passwords to be set via AuthLDAPBindPassword

Binds with empty passwords always succeed, but in case the password of the
user was not empty subsequent LDAP operations fail.


Before doing any bind check that the provided username is not NULL and that the
password is neither NULL nor empty.

Binds with empty passwords always succeed, but in case the password of the
user was not empty subsequent LDAP operations fail.
This causes authentications that use user supplied credentials
(AuthLDAPInitialBindAsUser set to on) to fail with status code 500 instead of
401 if the user supplied an empty password.


* Document r1885939 and r1885940

* Add lognumber

Reviewed by: rpluem, ylavic, covener

Github: closes #168


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/2.4.x@1887080 13f79535-47bb-0310-9956-ffa450edef68