mod_proxy: Bump shared worker name to 384 chars.  PR 53218.
This is a backport of r1896253 compatible with 2.4.x, where we can't change the
layout/size of the existing proxy_worker_shared fields, to allow for larger
worker name in <Proxy> and ProxyPass.

This backport adds a "name_ex" field of 384 chars at the end of the struct, the
existing "name" field is still initialized as before for third-parties that
might use it but every internal use is replaced by "name_ex".