mpm_event: Take advantage of the now fully non-blocking listener loop.
Now that the listener loop is fully non-blocking, we can ask for and use
only two timestamps accross loop: one for before poll()ing and the other
one for after poll()ing. After poll()ing the timestamp might drift a bit
while processing all the events, but if the queues maintenance is skipped
because of that it will happen in the next loop with timeout = 0.

Also process_timeout_queue() can be quite simplified if it does not need
to release the lock (all the callbacks are non-blocking too).

* server/mpm/event/event.c(TIMEOUT_EXPIRED):
  New helper to determine whether a timestamp + timeout expired, also
  expiring entries if the clock went backwart too much.

* server/mpm/event/event.c(listener_thread):
  Use two time1 and time2 variables for the time(stamp)s before and
  after poll()ing. time2 is fetched just after poll() and reused for all
  the below processing (i.e. not recomputed for queues maintenance).

* server/mpm/event/event.c(process_timeout_queue):
  Simplify the code by calling the callback in a single (double-)loop.
  The timeout_mutex mutex is not released/re-acquired anymore.
  Use TIMEOUT_EXPIRED() to check for expiry.